Files uploaded (oversight)
Just realized exactly why the hamburger icons were not appearing.  THEY DON'T EXIST.. :)
Wondering why on Earth they weren't appearing on the page.  Then it suddenly dawned on me that I haven't uploaded the images which I used to create the icon (one coloured black and the other blue) They are the images that I've used for 'list-style-image = url(" image.format")' in CSS.  It was working on my home computer obviously but it only just clicked that had only uploaded the original image (white) and not the others which I created later.