Update index.html
Added some 'white space' after the word 'default in the menu. Because the word is a list item, this will make use of the CSS property, white-space. 
So one you hover over that list item in the menu, not only will the background colour of the 'word' change, but also of the space after it; so the whole width of the that layer in the menu will highlight.