navigation buttons update
Buttons for left, right and rotate were sometimes not affecting the tetriminos.  Turns out that, because Font Awesome icons are nested in each button, and sometimes it was the icon press that was detected rather than the button press so there was no response.  I've not added the icon IDs to the switch statement so if I miss the button because I was exactly on an icon, the navigation function will still activate