blackjack and tetris changes
Wrote in a comment about changing if/else to 'switch' - actually done in another iteration of the game: the refactored one.

Added images nested in navigation and game state buttons to event listeners so if images are clicked they will have the same effect as the buttons they are nested in.