Update style.css
Changed how buttons respond to activation... now they are highlighted with a white offset outline.

I did this because there were potential problems with changing the size of buttons on activation given that the button sizes change depending on landscape or portrait functions - and that sizes are unpredictable on mobile devices; especially in landscape mode.