two fixes, hour 24 and tab refresh
Fixed issues related to midnight displaying as 24:00:00, now displays as 00:00:00
Also, since, when the user clicks to another tab and then returns to the clock, the numbers jumble up. I've added code to reset the clock if user has clicked on another tab and then returns to the clock tab.  The clock will display 00:00:00 for a second, and then show correct time with numbers displaying correctly