Basket complete
Basket functionality complete. So basket is saved prior to exiting the application.  So when returning to the app user can continue shopping from where they left off with the pick list re-loading with leftover list items and the basket being re-loaded with items that were dropped into the basket and saved prior to the page refreshing or closure of the page.

Number of items also implemented so that the user must select at least '1' of an item in the add to basket modal prior to being able to add the item to the basket, if no number is selected an alert modal will pop up telling user to select a number for how many items they want to drop into the basket.