added a CLEAR LIST button
Added a temporary 'clear list' button so that I can reset things if I want to re-create the list to test how newly implemented functions are working.  Maybe will keep the button and find another spot to put it rather than the main page; it might come in handy for errors.