checkout complete and shopping delivery allowed
Some changes; purchase complete button now active and so is delivery button.  Still some work to do dealing with which buttons to show if shopping has started but is incomplete or if list has started and is incomplete.  Most buttons  do not need to appear for most of the time on the main page; their presence causes confusion even though it's possible to have alerts come up for when user presses a button which cannot execute because it doesn't meet the conditions that allow it to do so.  Will work on this so that only 'currently needed' buttons display at any given time; reducing the potential for user confusion.