BUTTON ISSUES
Temporary shopping list now saved so that on returning to the app after refreshing or closing it, the shopping list is still available to be added to.

Some difficulties with displaying correct buttons at specific stages during list preparation or shopping or returning to app after closing mid shop or mid list prep.  Will work on this next as need to go through each stage in the list making / shopping process. I'll use entries in local storage to dictate which buttons should display based on what is currently in local storage; which itself is based on exactly where the user is in the process.