lots of minor issues worked on
Most major out of these is that if user saves the shopping list prior to uploading, and has to return to complete the list, once the list is uploaded. The saved items do not appear in the uploaded list. Only items added to the list that are created in the 'same' session as the upload were appearing in the uploaded shopping list.  This is now solved.  I took the data from the childnodes of the re-populated shopping list and put them in the array that is used to create the JSON file for upload.   Now the user can exit as many times as they wish because the old saved list items (and the currently added list items) will appear in the array used for the JSON file data.

Other major issue was buttons appearing at the wrong time and needed buttons not appearing, therefore halting the process.  It's an ongoing issue but largely fixed. I'm sure there will be some use cases that I haven't thought of.