Update script.js
Dealt with more incorrectly appearing buttons. This happens if when in the process of button presses, you decide to ignore the button press and refresh the page, then then the wrong buttons sometimes appear.   Largely resolved.  Only button giving slight trouble is the 'clear' button which clears local storage so that the app can restart in its default configuration with just the 'create list' button and clear buttons on the home screen.  Obviously, if the clear button is missing this will present issues - I can always clear local storage from inspect element in the browser; but if on the road that will be an issue so I need to resolve that one.  That's probably the last issue on the app - until I add new features such as the option to add categories, category items, images and descriptions.  Coming somewhere down the line.