removal of large images
removed all large images, most of them had dimensions of 1500x1500 and left smaller ones of size 300x300, most of which were around sizes of 30-40kb.

Did this because I'm planning on making the app available as a PWA, and caching most of the files since JS, CSS, HTML and the product images, will be needed when the app is offline, to avoid internet connection issues disabling parts of the app; which I experienced on first test when the phone forced the browser to go offline.

In the future I may consider chaching only the images that are relevant to the uploaded shopping list, which will reduce the size of the catch significantly.  If I introduce an 'add product' functionality, it can use a default image until there is an internet connection where user will be able to select an image of via some sort of search engine, perhaps one of those random image generators, but where you can have a category related image.