updates
added error page and image for page not found message.
event listener added to select element so that if a change is detected in selection, the selected country's data will be used to display its flag, name and capital city in the display div to the right of the screen.

The display div is made clickable and opens another tab that displays aa country map at google maps (or openStreetMap)