Day period style changes
Changed how the page looks depending on what time of day page is refreshed.  There are two modes; night and day, which change according to the sunrise and sunset times so that between sunrise and sunset there is a light mode and at night time, between sunset and sunrise there is a dark mode.

Background image added, and this also changes relevant to the time of day.

When a new day is clicked the hours scroll back to 00:00 on the clicked day so that user sees the weather from the start of the day.  When the current day is clicked, the hours will scroll back to the current hour and display 24 hours weather from current.