more updates
Refactor after implementation of search bar functionality. Search returns most popular cities (maximum of 5). Then, select element appears with the options.

when option is selected, latitude and longitude values of the city are used in the fetch URL to get the weather for the selected city.  A fetch is also made to get the city's time difference via its local time zone, and that is used to display the hour weather beginning with the closest local time hour of that city.

The screen also render's day mode or night mode according to the local time of day for that city; this was possible by using the time difference fetched.