adding product creation function
adding category creation and product creation to the application so that user can choose to add products to categories or create entirely new categories.

So far just added the ability to choose which to add, category or category product;

in next commit should have implemented the actual addition function.  After adding a category, user can re-enter from main page and then add a product to the category which will now be available (or it might be an idea to allow user to choose to add products immediately after adding a category)

Maybe there should be an option to add more categories or add more products immediately after submitting new categories/or products - that can be for another iteration if this current one works out.