MAJOR refactor
Managed to rewrite most functions and reduced eliminated around 600 lines of mainly code. It's the best I can do for the moment.  Currently learning Node JS to be enable 'create new categories and products' functionality.

The functions are already there; It's just a matter of learning how to save them to a JSON file so the new products and/or functions will be available once user reopens the app for another list creation session.

I may use local storage for the moment until I understand the node JS part of this.

OUTSTANDING ISSUES;
The new category function doesn't allow user to pick a product after product has been added to the category.  The category can be selected and the products will show, but attempting to submit the product to shopping list results in an error stating that no item has been selected.

Also, when user exits mid-shopping, but then reopens the app, although the remaining items not in the basket can be seen, they are no longer selectable and therefore cannot be pushed to the basket.