Update script.js
ISSUE #1 corrected.
When creating a new category it was not clickable if the newly created categoryName ( the key/value inside the category Object in JS) was more than one word, i.e. there was a space in the string. Solved using string.replace(' ', '-'), putting a hyphen between any multiple worded string that is used as category name in the category object in JS. The actual multi-worded category name on-screen has spaces because I don't use string.replace() for the name rendered to the screen;

ISSUE #2 - solved (but questions remain)
After adding a new product to a newly created category, the checkbox 'checked' was not registering on the new product and so an error was returned when trying to submit the visibly checked new product.

There seemed to be no difference between new category/ new product, and default categories/products when I checked the developer console.

I then tried to create a category, followed by two new products, and then the checked products were submitted.  The problem was that, for some reason, I had set an IF condition that, if on submit, the number of checkbox elements was greater than 1, then the product would be sent.  I changed that to >= 1, and that worked.  I think the issue might be that my code refers to the number of 'checked' products, rather than the number of products with checkboxes inside the form element.  Will have to investigate that further; I think I misinterpreted a log to the console which shows that, after submit,  the log showed; number of checked elements = elements + 1.

So if I selected 1 product, the log output was 2.  And that's why I was using the IF condition of:

if(elements checked > 1){
// execute code to process product
}