Update script.js
Added code to disable  'zoom on double-tap' which causes problems during gameplay.  There is no need for the zoom function during the game, but double-tapping (for example, if player is trying to rotate a tetrimino quickly to position it correctly) causes the screen to zoom into the tapped location, which would be one of the direction buttons or rotation buttons.

This causes the player to completely lose sight of what's happening on the gameplay screen, and usually results in a collision, or a tetrimino landing incorrectly positioned or incorrectly oriented while the player tries to un-zoom back to the normal screen.