Update script.js
for the geolocation city search I've changed the protocol from http to https since at the moment an error returns stating that the the request is being made from an insecure resource.

I'm not sure this will work and am looking at other solutions since the error also mentions 'mixed content', which I need to read up about. Will check once this update can be viewed and see if a further investigation into a solution is needed.