Update scriptb.js
Placeholder problem solved.
I was changing the value of the placeholder when an image was clicked because I wanted the previous dimensions to change. The placeholder dimensions don't need changing.

Setting the input element value to value=' ' clears the value and then the placeholder automatically appears

I've also reset the dimensions output to: textContent = 'W x H' in the same function, so that every time an image is clicked, output dimensions revert to
 the 'W x H' string indicating the output dimension parameters, and the inputs field for 'new width' and 'new height' are changed to the placeholders with the 'New Width' and 'New Height' strings.

ISSUES RESOLVED.

But instead