Update index.html
Ok, found the actual error.  Silly mistake really.  Didn't realize that the spelling of file names for switching back from GIF thumbnails to PNG thumbnails didn't comply with the case sensitivity, so the browser was not able to find the files which were named with title case.

Renamed files to lower case and changed HTML elements sources accordingly.  This should solve the issue.