Basket delete function added
Added a function that allows basket items to be deleted and returned to pick list.

Clicking on a basket item will open the 'amend-item' modal, where the clicked item can be removed from the basket and returned to the pick list.

Another option (currently being developed) allows user to amend the item count for the clicked item, with the lower limit being '1' item.

After confirming the change, the item will be returned to the basket with the updated count being reflected in the icon used to display item count.