Update script.js
Changed click events on basket products to a single click event on the basket.

After products leave the basket they are unaffected by click events on the basket.  Which has solved the issue of products moved from the basket still having events attached to them.  This was causing the 'amend' modal to open up if a product returned from the basket to the pick list was clicked in the pick list.

I will also make this change to the pick list so that, rather than the products having their own individual click events, the pick list element will have a single click event attached conditions can be used to process only those elements clicked that fit the 'product' criteria.

On the movement of products to other areas, those products will not have a click event attached and hopefully that will resolve conflicts where more than one modal is opening, or the wrong modal is opening after clicking a product because of events still attached to the product.