Update script.js
ISSUE: if checkout is viewed immediately after a basket item's count is amended, the new item amount is not recorded.

SOLVED - The checkout details were being taken from an array which had not been updated after the item count was amended.  Decided to use the stored array in localStorage for product names and count, and reject the use of the formerly used array