Update script.js
ISSUE: list button (which returns user to pick list) sometimes disappearing on entry into the checkout.

SOLVED: -
Noticed that had mistakenly chosen to 'hide' the list button itself on clicking it, rather than hiding the checkout;  The checkout was hidden immediately after anyway because that's what the 'closeModals' function does.

It appeared as though the list button was closing the checkout and opening the pick list, but what was actually happening is that the list button was hidden and then the closeModals function was closing the checkout, and then opening the pick list.

This happens so quickly that the disappearing the list button is not noticed.  Removed the line for hiding the button.  Now button always appears when the checkout is viewed.