Update script.js
ISSUE: on opening pick list, I noticed that some of the products had the delete icon showing on top of the plus icons (added prior to appending to pick list).

This isn't supposed to happen because in the function for uploading shopping list items, the delete image is removed from the product element.

ERROR FOUND:
Each time the shopping list is viewed , Newly selected items are displayed along with the previous items, and delete icons are added.  But, were being added to all items in the shopping list, so the previous items, already having the icon added, had one added again.  Each time the list was viewed, all previous product elements had the icon added, so icons were added multiple times if the list was viewed multiple times.

Only one of the icons were removed on upload and this is why there were still delete icons on some of the pick list products.

SOLVED: moved the image creation process so that the icon will be appended only to newly created shopping list products