Update script.js
Found a better way to render currently selected product amount into amend pick value element.

Instead of taking the value from the array which is quite long-winded, take it from the amount indicator of the product that is clicked, and put that value into the amend-pick-value field.

So the current amount of the product appears in that field and can be amended up or down (limits permitting).

NOTE* for a yet to be understood reason when a product amount is amended, if an attempt to amend again is made, with the old method of getting the amount from the array with amount numbers, the previous new number was not rendered into the amend-pick-value field; although the array is updated on confirming the changed number.  It's not the most important thing to know, but I'm curious to see what I missed.