item count issues solved
Amended basket items not showing amended product amount when returned to basket.

Some basket products were not showing the updated 'item amount', after the amount number was amended and product was returned to basket.  First noticed that products with one-word names worked as expected, displaying the correct amount after amendment and return to the basket, and some two-word product names were also displaying the correct amount.

On further examination it appeared that all product names of the form 'Product: type' were not rendering the item amount correctly.
The product names were being used as the amount icon ID by using productName.replace(' ', '-') to put a hyphen between the words of the product name to turn two worded names into a valid ID.  But this was causing an issue for product names with the 'product: type' format (example: 'quiche: veg') because productName.replace(' ', '-') didn't remove the colon, and a colon as part of an ID word is invalid; the id doesn't meet the requirements for CSS selectors.  CSS will not select the element because it has an invalid ID.  And that's why the new amount was not being rendered to the element.

Added the below method to a temporary id:
let tempID =  productName.replace(':', ''), to remove the colon, leaving no space.
(example  - 'quiche: veg' becomes 'quiche veg')

then used the below method for the final id:
let iconID =  productName.replace(' ', '-'), to replace the space separating the two words with a hyphen
(completed example  -  'quiche veg' becomes 'quiche-veg')