Category/product creation implemented
It's now possible to create and save new categories and/or new products to existing categories.  Newly saved categories/products are saved in a database in indexedDB.  The additional categories/products created by the user will populate the categories modal when the app is restarted after closure.

When categories and products are created when list building they are immediately available as they are instantly added to the existing, categories/products as temporary elements.  They will become semi-permanent elements once the app is restarted after closure (as long as they are still available in the indexedDB store).

In future iterations I'm planning on using a service worker to make added categories fully permanent and will add the option to delete non-default categories and/or products.