Create git mistake.txt
Checking if removing the git subdirectory from the apps main folder will solve the issue of the pages not deploying.  Stumbled across this by accident when using git bash to create a new repository and, adding this app to that other repository, got an error message stating that 'outer repository will not contain the contents of the embedded repository'.  That's when I realized that there was one repository inside the app's folder.