added some API call features
Added OAuth2 information for API calls to youtube.

I will not be using these for the moment but want to keep them for when I overhall the application to include API calls to youtube for subscription data and API calls for deletion of youtube subscriptions.

These features are already available in another application which I have built and deployed; a simpler application.  But I would like to flesh out that application with the menu data specific to the user when they are signed into google and using the application.