adding another iteration of solitair.
This iteration is far from complete but I'm adding an auto-solve system that activates when, at a certain point in the game, it is recognized that a solution is possible.  The game will then give the player the option to auto complete.

Currently working on a method to keep track of the card positions as the game progresses.  There is an easier way to do this, and that's to just find the position of each card at the moment when it is recognized that the game must have a solution; I've tried that and am having some difficulty because (probably due to the way I coded) sometimes the information about card positions is incorrect, and sometimes it is correct.

When positions are incorrect it either leads to a crash or an infinite loop.

So I decided to keep a contemporary record of the card movements.  There's a side advantage of this method, which is that it probably will make it possible to have a back button system to reverse moves, even all the way back to the beginning of the game.

That added possibility made me decide to continue on with the method of tracking card movements and positions while the game is in progress.

The code will need re-factoring because everything is all over the place at the moment and doesn't make much logical sense even though I tried to make the function names self explanatory. The first iteration was about a year ago, so hopefully I'll be able to improve on the codebase.