solved an issue where cards moved from drop piles to foundation piles were not being tracked.  The issue occurred because I was using a condition to decide what to do with the card based on its origin pile; but instead of using the value for the origin pile, I was mistakenly using the value for the destination pile.
This meant that the origin property on the card object NEVER matched the consitional value, because the origin was a drop pile, but the condition was looking for the property value of foundation pile (because I had accedentally set THAT as the origin).  I forgot that the 'event target' was the destination of the dropped card and NOT the origin. I was using the event target as the 'origin' condition to be met.

So the object was never pushed to the tracker because its origin was the drop pile, but the origin which was looked for in the condition was foundation pile.

This caused a side effect; cards which were moved from drop pile to drop pile had no issues because their origin IS their destination so the card object passed the mistaken condition and was still sent to the tracking function.

That's why I only noticed the error when transferring a card from drop pile to destination pile.  I think this is working fine now so I can move onto updating the tracking arrays for multiple moved cards