objects in tracking arrays (and in breadcrumb array for history) were showing incorrect values if the card was moved more than once.  This is because, although I changed the card destination property on the moved card, I didn't change the origin; this origin represents the initial origin, where the card sat immediately after distribution.  The true origin of a moved card is the previous destination of the card.
Solution: if the dropped card destination property had a value, I created a variable and assigned that value to it, then I set the variable to the value of the origin property. I then updated the destination property on the object.

If the dropped card had no value in the destination property, then the card had not been moved before, because when objects are created at card distribution, the destination property is left empty, since the card has not yet moved; so for dropped cards having an empty destionation property, for uniformity I set the origin property to the property value that was already on the object.