several changes:
resolved issue of drop cards which were rejected still appearing in the breadcrumb array as though they had been moved; that would have given an incorrect version of the movement history.

I created a temporary to hold the dragged card, and then, only if and only if the card drop was accepted, then the temporary card would be pushed to the breadcrumb array.

Also, figured out another issue where empty values were being pushed to the breadcrumb array.  This occurred because, after a card from the temporary array is pushed to the breadcrumb array, the temporary array is emptied.

But, when a waste pile card is dropped to foundation or a drop pile, there was an attempt to push whatever was in the temporary array to the breadcrumb array. Waste pile cards do not have a temporary tracking object. Their object is produced when it is recognized that the dropped card came from the waste pile.  So when they drop, there is no temporary tracking object and so an empty value from the tracking array was getting pushed to the breadcrumb array.

The cards that are dropped that originated in the waste pile get their own code to create a tracking object and that object is pushed directly to breadcrumb, bypassing the temporary array. This is so because waste pile cards have no real origin position 'until' they are dropped, so no object exists to track.  Other cards, from foundation pile and drop pile are given a default object just after distribution. And that is the object that is temporarily stored in the temp array pending a successful drop; if the drop is not successful then nothing is pushed to the breadcrumb array.

So in the functions that, just before sending the dropped card to be appended to a pile, I used the condition that if the temporary array has a value at index zero, then push it to the breadcrumb array.  This prevented empty values getting pushed to the array, and skewing the order of events for the movement history.

Waste pile cards would get appended automatically on meeting the conditions for a successful drop; After the drop happens, the function which assesses the origin of the dropped card then recognizes that the card originated in the waste pile and creates a tracking object for it, pushing it to the breadcrumb array.  Previously both the empty temporary array value, and the newly created waste pile tracking objects were being  pushed.  I'm not sure I actually commited that. Perhaps not, but I'm giving an explanation here so that there's an understanding of why the code has changed to the current state.